moved everything relatd to building to build_tools/ from root
